# chore: Minor rego optimization by removing excessive queries