--- a/enterprise/coderd/templates_test.go
+++ b/enterprise/coderd/templates_test.go
@@ -228,7 +228,6 @@ func TestTemplateACL(t *testing.T) {
 			Role:  codersdk.TemplateRoleView,
 		})
 	})
-
 }
 
 func TestUpdateTemplateACL(t *testing.T) {