# What this does

Removes this "always false" query in the list. Also enforces you must also be an org member for group acl to work.

```
input.object.org_owner = ""
input.object.org_owner != ""
```

# Before

```
cpu: Intel(R) Core(TM) i7-6770HQ CPU @ 2.60GHz
BenchmarkRBACFilter/NoRoles-8              11394            108734 ns/op           27871 B/op        565 allocs/op
BenchmarkRBACFilter/Admin-8              3897625               898.6 ns/op           385 B/op          0 allocs/op
BenchmarkRBACFilter/OrgAdmin-8             12157             90951 ns/op           28912 B/op        586 allocs/op
BenchmarkRBACFilter/OrgMember-8            10772            106122 ns/op           31898 B/op        641 allocs/op
BenchmarkRBACFilter/ManyRoles-8            26516             40330 ns/op           13193 B/op        292 allocs/op
BenchmarkRBACFilter/AdminWithScope-8      749647              1510 ns/op             139 B/op          3 allocs/op
```

```
+---------+--------------------------------------------------------------------+
| Query 1 | input.object.org_owner != ""                                       |
|         | input.object.org_owner in {"19a3d781-9e53-4608-948e-ac084470cd99"} |
|         | input.object.owner != ""                                           |
|         | "me" = input.object.owner                                          |
+---------+--------------------------------------------------------------------+
| Query 2 | input.object.org_owner = ""                                        |
|         | input.object.owner != ""                                           |
|         | "me" = input.object.owner                                          |
+---------+--------------------------------------------------------------------+
| Query 3 | input.object.org_owner != ""                                       |
|         | input.object.org_owner in {"19a3d781-9e53-4608-948e-ac084470cd99"} |
|         | input.object.org_owner != ""                                       |
|         | "read" in input.object.acl_group_list[input.object.org_owner]      |
+---------+--------------------------------------------------------------------+
| Query 4 | input.object.org_owner != ""                                       |
|         | input.object.org_owner in {"19a3d781-9e53-4608-948e-ac084470cd99"} |
|         | input.object.org_owner != ""                                       |
|         | "*" in input.object.acl_group_list[input.object.org_owner]         |
+---------+--------------------------------------------------------------------+
| Query 5 | input.object.org_owner = ""                                        |
|         | input.object.org_owner != ""                                       |
|         | "read" in input.object.acl_group_list[input.object.org_owner]      |
+---------+--------------------------------------------------------------------+
| Query 6 | input.object.org_owner = ""                                        |
|         | input.object.org_owner != ""                                       |
|         | "*" in input.object.acl_group_list[input.object.org_owner]         |
+---------+--------------------------------------------------------------------+
| Query 7 | "read" in input.object.acl_user_list.me                            |
+---------+--------------------------------------------------------------------+
| Query 8 | "*" in input.object.acl_user_list.me                               |
+---------+--------------------------------------------------------------------+
```

# After

```
cpu: Intel(R) Core(TM) i7-6770HQ CPU @ 2.60GHz
BenchmarkRBACFilter/NoRoles-8              16071            101915 ns/op           21373 B/op        442 allocs/op
BenchmarkRBACFilter/Admin-8              3791935               749.4 ns/op           396 B/op          0 allocs/op
BenchmarkRBACFilter/OrgAdmin-8             14983             81771 ns/op           24588 B/op        503 allocs/op
BenchmarkRBACFilter/OrgMember-8            12266             90102 ns/op           26505 B/op        539 allocs/op
BenchmarkRBACFilter/ManyRoles-8            27688             38889 ns/op           13237 B/op        291 allocs/op
BenchmarkRBACFilter/AdminWithScope-8      761229              1480 ns/op             139 B/op          3 allocs/op
```

```
+---------+--------------------------------------------------------------------+
| Query 1 | input.object.org_owner != ""                                       |
|         | input.object.org_owner in {"19a3d781-9e53-4608-948e-ac084470cd99"} |
|         | input.object.owner != ""                                           |
|         | "me" = input.object.owner                                          |
+---------+--------------------------------------------------------------------+
| Query 2 | input.object.org_owner = ""                                        |
|         | input.object.owner != ""                                           |
|         | "me" = input.object.owner                                          |
+---------+--------------------------------------------------------------------+
| Query 3 | input.object.org_owner != ""                                       |
|         | input.object.org_owner in {"19a3d781-9e53-4608-948e-ac084470cd99"} |
|         | "read" in input.object.acl_group_list[input.object.org_owner]      |
+---------+--------------------------------------------------------------------+
| Query 4 | input.object.org_owner != ""                                       |
|         | input.object.org_owner in {"19a3d781-9e53-4608-948e-ac084470cd99"} |
|         | "*" in input.object.acl_group_list[input.object.org_owner]         |
+---------+--------------------------------------------------------------------+
| Query 5 | "read" in input.object.acl_user_list.me                            |
+---------+--------------------------------------------------------------------+
| Query 6 | "*" in input.object.acl_user_list.me                               |
+---------+--------------------------------------------------------------------+
```